modify jobs controller

+++ controllers/jobs.js
@@ -67,6 +67,13 @@ exports.updateJob = async (req, res) => {
 exports.deleteJob = async (req, res) => {
   const { id: jobID } = req.params
   const { userID } = req.user
+  const { company, position } = req.body
+
+  if (!company || !position) {
+    return res
+      .status(StatusCodes.BAD_REQUEST)
+      .json({ msg: 'Please provide company and position' })
+  }
   try {
     const job = await Jobs.findByIdAndRemove({ _id: jobID, createdBy: userID })
     if (!job) {
